+++ src/functions/audit/mod.rs
@@ -1,0 +1,10 @@
+pub mod auditCount;
+pub mod auditEntries;
+pub mod auditEntryAction;
+pub mod auditEntryChanges;
+pub mod auditEntryID;
+pub mod auditEntryReason;
+pub mod auditEntryTarget;
+pub mod auditEntryUser;
+pub mod auditLatest;
+pub mod helpers;